+++ app/pilot/page.tsx
@@ -1,0 +1,5 @@
+import { PilotAnalyticsDashboard } from "@/components/revision/PilotAnalyticsDashboard";
+
+export default function PilotPage() {
+  return <PilotAnalyticsDashboard />;
+}